requirements.txt
numpy==1.17.4
pytorch==1.1.0
torchvision==0.3.0
six==1.13.0
h5py==2.10.0
pillow==5.2.0
scipy==1.1.0
scikit-learn==0.22.2
metric-learn==0.5.0